+++ src-tauri/src/lib.rs
@@ -61,10 +61,10 @@ pub fn run() {
             workstation::data::excel::upload_excel_file,
             workstation::data::excel::get_sheet,
             sentiment_analysis::handler::set_sentiment_analysis_target_column,
-            sentiment_analysis::handler::calculate_sentiment_analysis_indonesia,
-            sentiment_analysis::handler::calculate_sentiment_analysis_multilanguage,
-            sentiment_analysis::handler::calculate_sentiment_analysis_english,
-            sentiment_analysis::handler::analyze_and_update_sentiment,
+            sentiment_analysis::calculate::calculate_sentiment_analysis_indonesia,
+            sentiment_analysis::calculate::calculate_sentiment_analysis_multilanguage,
+            sentiment_analysis::calculate::calculate_sentiment_analysis_english,
+            sentiment_analysis::calculate::analyze_and_update_sentiment,
             social_network::handler::set_vertices,
             social_network::handler::calculate_centrality,
             database::lib::handler::load_sqlite_data,
