+++ src-tauri/src/workstation/sentiment_analysis/mod.rs
@@ -1,2 +1,3 @@
 pub mod state;
 pub mod handler;
+pub mod calculate;
